Fix for scopebuddy.py not finding config.json
Hi, great project! I was running into the following error: `Config file doesn't exist. A sample file is included in the repository for this project. [Errno 2] No such file or directory: '/config.json`. My take is that the code you had for getting the current working directory wasn't working for some reason. Luckily, removing it allows the script to find the config.json file in the current working directory, so it was an easy fix.